refactor flink stream-stream join rules
