+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/rules/physical/stream/StreamExecTemporalJoinRule.scala
@@ -18,14 +18,10 @@
 
 package org.apache.flink.table.planner.plan.rules.physical.stream
 
-import java.util
-
-import org.apache.flink.table.planner.plan.`trait`.FlinkRelDistribution
-import org.apache.flink.table.planner.plan.nodes.FlinkConventions
+import org.apache.flink.table.planner.plan.nodes.FlinkRelNode
 import org.apache.flink.table.planner.plan.nodes.logical._
 import org.apache.flink.table.planner.plan.nodes.physical.stream.StreamExecTemporalJoin
-import org.apache.flink.table.planner.plan.utils.{FlinkRelOptUtil, IntervalJoinUtil, TemporalJoinUtil}
-import org.apache.calcite.plan.RelOptRule.{any, operand}
+import org.apache.flink.table.planner.plan.utils.TemporalJoinUtil
 import org.apache.calcite.plan.{RelOptRule, RelOptRuleCall, RelTraitSet}
 import org.apache.calcite.rel.RelNode
 import org.apache.calcite.rel.core.JoinRelType
@@ -36,11 +32,7 @@ import org.apache.flink.util.Preconditions.checkState
  * the temporal join node is a [[FlinkLogicalJoin]] which contains [[TemporalJoinCondition]].
  */
 class StreamExecTemporalJoinRule
-  extends RelOptRule(
-    operand(classOf[FlinkLogicalJoin],
-      operand(classOf[FlinkLogicalRel], any()),
-      operand(classOf[FlinkLogicalRel], any())),
-    "StreamExecTemporalJoinRule") {
+  extends StreamExecJoinRuleBase("StreamExecJoinRuleBase") {
 
   override def matches(call: RelOptRuleCall): Boolean = {
     val join = call.rel[FlinkLogicalJoin](0)
@@ -60,59 +52,22 @@ class StreamExecTemporalJoinRule
   }
 
   private def matchesTemporalTableFunctionJoin(join: FlinkLogicalJoin): Boolean = {
-    val joinInfo = join.analyzeCondition
-    val tableConfig = FlinkRelOptUtil.getTableConfigFromContext(join)
-    val (windowBounds, _) = IntervalJoinUtil.extractWindowBoundsFromPredicate(
-      joinInfo.getRemaining(join.getCluster.getRexBuilder),
-      join.getLeft.getRowType.getFieldCount,
-      join.getRowType,
-      join.getCluster.getRexBuilder,
-      tableConfig)
+    val (windowBounds, _) = extractWindowBounds(join)
     windowBounds.isEmpty && join.getJoinType == JoinRelType.INNER
   }
 
-  override def onMatch(call: RelOptRuleCall): Unit = {
-    val join = call.rel[FlinkLogicalJoin](0)
-    val left = call.rel[FlinkLogicalRel](1)
-    val right = call.rel[FlinkLogicalRel](2)
-
-    val newRight = right match {
-      case snapshot: FlinkLogicalSnapshot =>
-        snapshot.getInput
-      case rel: FlinkLogicalRel => rel
-    }
-
-    def toHashTraitByColumns(
-        columns: util.Collection[_ <: Number],
-        inputTraitSets: RelTraitSet) = {
-      val distribution = if (columns.size() == 0) {
-        FlinkRelDistribution.SINGLETON
-      } else {
-        FlinkRelDistribution.hash(columns)
-      }
-      inputTraitSets.
-        replace(FlinkConventions.STREAM_PHYSICAL).
-        replace(distribution)
-    }
-
-    val joinInfo = join.analyzeCondition
-    val (leftRequiredTrait, rightRequiredTrait) = (
-      toHashTraitByColumns(joinInfo.leftKeys, left.getTraitSet),
-      toHashTraitByColumns(joinInfo.rightKeys, newRight.getTraitSet))
-
-    val convLeft: RelNode = RelOptRule.convert(left, leftRequiredTrait)
-    val convRight: RelNode = RelOptRule.convert(newRight, rightRequiredTrait)
-    val providedTraitSet: RelTraitSet = join.getTraitSet.replace(FlinkConventions.STREAM_PHYSICAL)
-
-    val temporalJoin = new StreamExecTemporalJoin(
+  override protected def transform(
+      join: FlinkLogicalJoin,
+      convertedLeft: RelNode,
+      convertedRight: RelNode,
+      providedTraitSet: RelTraitSet): FlinkRelNode = {
+    new StreamExecTemporalJoin(
       join.getCluster,
       providedTraitSet,
-      convLeft,
-      convRight,
+      convertedLeft,
+      convertedRight,
       join.getCondition,
       join.getJoinType)
-
-    call.transformTo(temporalJoin)
   }
 }
